// packages/ml-ops/src/types.d.ts
export declare const TrainingConfigSchema: import("@sinclair/typebox").TObject<{
    method: import("@sinclair/typebox").TUnion<[import("@sinclair/typebox").TLiteral<"qlora">, import("@sinclair/typebox").TLiteral<"ternary-bitnet">, import("@sinclair/typebox").TLiteral<"grpo">, import("@sinclair/typebox").TLiteral<"full-finetune">, import("@sinclair/typebox").TLiteral<"lora">]>;
    baseModel: import("@sinclair/typebox").TString;
    outputName: import("@sinclair/typebox").TString;
    outputDir: import("@sinclair/typebox").TString;
    dataset: import("@sinclair/typebox").TString;
    datasetSplit: import("@sinclair/typebox").TString;
    epochs: import("@sinclair/typebox").TNumber;
    batchSize: import("@sinclair/typebox").TNumber;
    gradientAccumulationSteps: import("@sinclair/typebox").TNumber;
    learningRate: import("@sinclair/typebox").TNumber;
    weightDecay: import("@sinclair/typebox").TNumber;
    warmupSteps: import("@sinclair/typebox").TNumber;
    maxSeqLength: import("@sinclair/typebox").TNumber;
    precision: import("@sinclair/typebox").TUnion<[import("@sinclair/typebox").TLiteral<"fp16">, import("@sinclair/typebox").TLiteral<"bf16">, import("@sinclair/typebox").TLiteral<"fp32">, import("@sinclair/typebox").TLiteral<"8bit">, import("@sinclair/typebox").TLiteral<"4bit">]>;
    loraRank: import("@sinclair/typebox").TNumber;
    loraAlpha: import("@sinclair/typebox").TNumber;
    loraDropout: import("@sinclair/typebox").TNumber;
    grpoBeta: import("@sinclair/typebox").TOptional<import("@sinclair/typebox").TNumber>;
    grpoRewardModel: import("@sinclair/typebox").TOptional<import("@sinclair/typebox").TString>;
    saveSteps: import("@sinclair/typebox").TNumber;
    evalSteps: import("@sinclair/typebox").TNumber;
    loggingSteps: import("@sinclair/typebox").TNumber;
    useUnsloth: import("@sinclair/typebox").TBoolean;
    useFlashAttention: import("@sinclair/typebox").TBoolean;
    deepspeed: import("@sinclair/typebox").TOptional<import("@sinclair/typebox").TString>;
    fsdp: import("@sinclair/typebox").TBoolean;
    resumeFromCheckpoint: import("@sinclair/typebox").TOptional<import("@sinclair/typebox").TString>;
    gpuIds: import("@sinclair/typebox").TString;
    seed: import("@sinclair/typebox").TNumber;
}>;
export type TrainingConfig = Static<typeof TrainingConfigSchema>;
// ...
export declare const DEFAULT_TRAINING_CONFIG: TrainingConfig;